requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.10.0

# Azure Services
azure-search-documents>=11.4.0
azure-ai-formrecognizer>=3.3.0
azure-identity>=1.15.0

# Vector Store (Local Fallback)
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.3.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.12.0
tiktoken>=0.5.2

# Web Framework
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Evaluation
ragas>=0.1.0
datasets>=2.16.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
tqdm>=4.66.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0
